feat(ODIN C4): the count currency is the CHIP's declared word, and the wide MLP exports
THE WALL C3 MEASURED. `t0_54`'s simple_mlp deploys 0.9800 on the wide fabric
and was refused at EXPORT every time by a universal 127 — readout emission
bound 146. C2 had already proved 127 is not fabric (the generated core carries
no count field at all; multiplicity is k adjacent AER transactions), so it was a
SOFTWARE representation shared by four implementations, and a representation is
a thing a chip gets to declare.

ONE SSOT. `count_ceiling(chip_claims)` in `models/spiking/serial/refusals.py`:
the largest positive value a SIGNED word of the width the chip declares its
registers at holds, floored at a byte and capped by the currency's own word.
That width is the one every implementation instantiates the chip's law at —
nevresim's `EventSerialIntegrate<bits>`, the generated core's accumulator, the
exporter's theta ceiling — so it is a claim and not an analogy.

  membrane_bits  8 -> 127     the STOCK fabric, untouched, byte for byte
  membrane_bits 16 -> 32,767  the WIDE fabric, which is what its arms carry

Every holder now consumes it and none carries a literal: the torch fold
(`count_ceiling(soma_law)`), the stock exporter and its manifest, the generated
fabric's gate (`count_ceiling(spec)`), and the envelope auditor — which reads
the ceiling off the law the SHIPPED BYTES declare, so a bundle cannot be audited
against a currency its own chip never claimed. `EMISSION_CEILING`'s duplicate
literal is gone; it is the SSOT's default point now. The currency is still
ASSERTED and never clamped — only the threshold moved.

NEVRESIM (submodule 482213c, branch odin-integration-policy). `spike_t` was
`int_fast8_t`; it is `std::int16_t` — exactly 16 bits everywhere, because a
deployment's counts must not depend on how a compiler sizes a "fast" type. The
counted raster's store followed it (`SpikeTrainRecorder` held counts in a
`uint8_t` and would have wrapped 300 to 44 in silence), and
`k_emitted_spike_ceiling`'s literal 127 became a per-instantiation
`count_currency_ceiling(MembraneBits)` computing the same number as the Python
SSOT. Host side: the SPKTRN2 parser now guards the WORD (a corrupt line) rather
than one chip's ceiling, and the re-packed rasters are `int16`.

PARITY, WHICH IS THE TRUTH HERE. hw runner 143 -> all green at ZERO count
movement: every cosim, parity and certificate gate that existed still reports
the same numbers. Four mutants die — pinning the C++ ceiling back to 127 fails
the static_assert, narrowing `spike_t` fails to compile the case at all,
`count_ceiling` collapsed to 127 kills 7 tests and to 32,767 kills 15.

THE NEW GATE. A synthetic high-multiplicity case on a 16-bit chip: 300 spikes
from one neuron in one cycle — past 127 AND past an unsigned byte, so no store
survives it by luck — proved EXACT on three independent arms (the torch fold,
the generated RTL through a cosimulation that emits 300 adjacent AER
transactions, and nevresim through both SPKREC and the counted SPKTRN2 raster),
with the same events refused by name on an 8-bit register.

TWO DEFECTS THE PAYOFF RUN EXPOSED, both invisible while every bundle was
two-hop and index-ordered:

  * `pass_builds` emitted `range(len(cores))`. The MLP's cascade is 1 -> 2 -> 0,
    so the shipped executor refused the bundle by name — a readout scheduled
    before the core it reads. `causal_core_order` is now a topological sort with
    a deterministic tie-break and a cycle refusal, and the freezer emits it.
  * the envelope auditor rebuilt the network in PLAN-LIST position. A pass order
    is a schedule, not an addressing scheme: cores are placed at the index their
    routes name, and two plans claiming one index are a finding.

Suite 13,716 passed / 33 skipped / 0 failed, typecheck 0, budgets 0,
undefined-names 0, hw runner green, package harness 143 gates / 0 red.

+++ src/mimarsinan/chip_simulation/nevresim/instrumented_builds.py
@@ -149,10 +149,11 @@ def predict_spiking_raw_with_spike_trains(
         record_spike_trains=True,
         timeout_s=driver.simulation_step_timeout_s,
     )
-    # uint8 holds the whole count currency (ceiling 127) as well as the bit
-    # raster, so both wire versions land in the same array type.
+    # int16 IS the count currency's word (nevresim's ``spike_t``), so it holds
+    # any chip's counts as well as the bit raster and both wire versions land
+    # in the same array type.
     as_arrays = [
-        {core: np.array(trains, dtype=np.uint8).reshape(len(trains), -1)
+        {core: np.array(trains, dtype=np.int16).reshape(len(trains), -1)
          for core, trains in sample.items()}
         for sample in spike_trains
     ]
